Simulation Debugging
Code now works for a single simulation.  Some debugging statements are added to track how the turtle goes around the grid.  The gene creation function was also amended so that the creation of genes is much more flexible (ie. it is possible to change the number of direction and action pairs, if needed).